[LEMS-4255/Radio-regression-stories] Trimmed more unnecessary tests according to talks with Mark and Sarah P. (Specifically, return the single select with rationale to initial state, removed choose 3, and converted many choices test to a story rather than a test
