+++ packages/perseus/src/widgets/radio/__docs__/multiple-choice-interactions-regression.stories.tsx
@@ -133,45 +133,6 @@ export const FocusMultiSelect: Story = {
     },
 };
 
-export const SingleSelectWithRationale = {
-    decorators: [radioRendererDecoratorWithDebugUI],
-    args: {
-        choices: [
-            generateRadioChoice("USS Voyager (NCC-74656)", {
-                rationale: "Commanded by Captain Kathryn Janeway.",
-            }),
-            generateRadioChoice("USS Enterprise (NCC-1701)", {
-                rationale:
-                    "\nThis rationale has a blank line at the start, which should **NOT** affect the rendered rationale. More text: " +
-                    "Shields up. I recommend we transfer power to phasers and arm the photon torpedoes. Something strange on the detector circuit. " +
-                    "The weapons must have disrupted our communicators. You saw something as tasty as meat, but inorganically materialized out of patterns used by our transporters. " +
-                    "Captain, the most elementary and valuable statement in science, the beginning of wisdom, is 'I do not know.'" +
-                    '\n\n**Top tip!** This is the ship he commands in the series, but it is not his first command. Watch *"The Battle"* (Season 1, Episode 9) for more. And, as always, beware of Ferengi!',
-            }),
-            generateRadioChoice("USS Enterprise (NX-01)", {
-                rationale: "Commanded by Captain Jonathan Archer.",
-            }),
-            generateRadioChoice("USS Stargazer (NCC-2893)", {
-                correct: true,
-                rationale:
-                    "**This is the correct choice.** In one of the battles with the Ferengi, he killed the son of DaiMon Bok, who later sought revenge on Picard.",
-            }),
-        ],
-    },
-    play: async ({canvas, userEvent}) => {
-        const choiceToClick = canvas.getByRole("button", {
-            name: /^\(Choice D\)/,
-        });
-        await userEvent.click(choiceToClick);
-
-        const checkAnswerButton = canvas.getAllByRole("button", {
-            name: "Check answer",
-        })[0];
-        await userEvent.click(checkAnswerButton);
-        await checkAnswerButton.blur();
-    },
-};
-
 export const SelectChoiceMoveFocusAfter: Story = {
     decorators: [radioRendererDecoratorWithDebugUI],
     args: {
